Adding preliminary support for web-servers
For any real-world CKAN implementation running the CKAN python application through a web-server is a fundamental requirement.

CKAN recommends using Apache as the web-server to interface with the the CKAN application, with Nginx running over the top to provide caching support.

Support for both these web-servers is now possible from this commit forwards. Currently many of these changes are out of scope of this module and therefore many of these changes will effectively be moved from this role to more suitable locations. This commit marks the start of this refactoring process.