Removing duplicate task for adding app user to www-data group and clarifying other tasks
Since CKAN uses a non-standard web root (i.e. not /var/www) ownership needs to be granted for the www-data group to have access. This should be done by this role, rather than the apache role as it is this role that is doing something non-default and should therefore do any required work to ensure Apache can operate as normal.